visualization/requirements.txt
dash
dash_bootstrap_components
hydra-core
joblib
pandas
pygments
sshtunnel_requests
